kms: move EncryptionConfigurationComputer interface to its own file
Extract the EncryptionConfigurationComputer interface and
NoopEncryptionConfigurationComputer placeholder from the KMS preflight
controller file into encryption_configuration_computer.go.
